requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
solana==0.29.0
snscrape==0.6.2.20230320
textblob==0.17.1
psycopg2-binary==2.9.9
solders
